requirements.txt
pyyaml>=6.0.1
openai>=1.40.0